+++ src/main/java/day5seeds/SeedMapping.java
@@ -3,12 +3,12 @@ package day5seeds;
 import java.util.List;
 import java.util.Map;
 
-public record SeedMapping(List<Integer> seeds,
-                          Map<Integer, Integer> seedToSoil,
-                          Map<Integer, Integer> soilToFertilizer,
-                          Map<Integer, Integer> fertilizerToWater,
-                          Map<Integer, Integer> waterToLight,
-                          Map<Integer, Integer> lightToTemperature,
-                          Map<Integer, Integer> temperatureToHumidity,
-                          Map<Integer, Integer> humidityToLocation) {
+public record SeedMapping(List<Long> seeds,
+                          Map<Long, Long> seedToSoil,
+                          Map<Long, Long> soilToFertilizer,
+                          Map<Long, Long> fertilizerToWater,
+                          Map<Long, Long> waterToLight,
+                          Map<Long, Long> lightToTemperature,
+                          Map<Long, Long> temperatureToHumidity,
+                          Map<Long, Long> humidityToLocation) {
 }
